+++ ac/CodeAnalysis/ExpressionSyntax.cs
@@ -2,6 +2,6 @@ namespace Aromat.CodeAnalysis
 {
     abstract class ExpressionSyntax : SyntaxNode
     {
-        public abstract int Calculate();
+        public abstract decimal Calculate();
     }
 }
